requirements.txt
fastapi
uvicorn
rich
notion_client
python-dotenv